backend/requirements.txt
# FastAPI and web server
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.9
email-validator>=2.1.0  # For Pydantic's EmailStr

# Database (choose one)
# SQLAlchemy>=2.0.27
# pymongo>=4.6.1

# NLP and ML libraries
spacy>=3.7.2
scikit-learn>=1.4.0
numpy>=1.26.3
pandas>=2.2.0

# Speech-to-Text
# openai>=1.14.0 # For Whisper API

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.1
requests>=2.31.0
pytest>=8.0.0  # For testing

# For production
# gunicorn>=21.2.0  # For production deployment
# sentry-sdk>=1.39.2  # For error tracking 